requirements.txt
asgiref==3.3.2
Django~=3.2.3
djangorestframework<3.12
drf-yasg==1.17.1
Pillow==8.0.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.4.1
uuid==1.30
